Add files via upload
There are N plants in a garden. Each of these plants has been added with some amount of pesticide. After each day, if any plant has more pesticide than the plant on its left, being weaker than the left one, it dies. You are given the initial values of the pesticide in each plant.
This algorithm prints the number of days after which no plant dies, i.e. the time after which there are no plants with more pesticide content than the plant to their left.
For more information:
https://www.hackerrank.com/challenges/poisonous-plants/problem